backend/requirements.txt
fastapi
uvicorn
openai
anthropic
google-generativeai
python-dotenv
requests
tenacity